more code editor hooks
