+++ src/shared/ui/CodeEditor/hooks/index.ts
@@ -1,2 +1,8 @@
-export {useMonacoSetup, type SupportedLanguage} from "./useMonacoSetup"
-export {useDecorations, type HighlightGroup, type HighlightRange} from "./useDecorations"
+export {useMonacoSetup} from "./useMonacoSetup"
+export {useDecorations} from "./useDecorations"
+export {useEditorEvents} from "./useEditorEvents"
+export {useLanguageProviders} from "./useLanguageProviders"
+export {useFolding} from "./useFolding"
+
+export type {SupportedLanguage} from "./useMonacoSetup"
+export type {HighlightGroup, HighlightRange} from "./useDecorations"
